add asm to stack

+++ emulator/models/evm/simulator/evm_sim.h
@@ -77,7 +77,7 @@ public:
   PD_REG< uint64 >  cur_addr;
   PD_REG< uint64 >  caller_addr;
   PD_REG< uint64 >  origin_addr;
-  PD_REG< uint32 >  gas_price;
+  PD_REG< uint64 >  gas_price;
   PD_REG< uint64 >  ext_code_size;
   PD_REG< uint64 >  inp_data_size;
   PD_REG<uint64>  input_data[1024];
@@ -86,7 +86,7 @@ public:
   uint32  sp;
   uint1  is_pc_within_inst;
   uint1  is_pc_const_changed;
-  uint32  gas_available;
+  uint64  gas_available;
   uint32  mem_size;
   uint8  data_bus[1024];
   uint64  stack_arr[2048];
